ee838_video/hw3/source/requirements.txt
tensorflow-gpu
matplotlib
tqdm
pillow
scipy
numpy
imageio
animeface
pillow
scikit-image
opencv-python
